connecting QAD and system finished

+++ database/migrations/2024_12_09_021451_create_items_table.php
@@ -20,8 +20,8 @@ return new class extends Migration
             $table->foreignId('category_id')->constrained('categories');
             $table->timestamp('service_date')->nullable();
             $table->boolean('isDisposition')->nullable();
-            $table->integer('cost')->nullable();
-            $table->integer('nbv')->nullable();
+            $table->unsignedBigInteger('cost')->nullable();
+            $table->unsignedBigInteger('nbv')->nullable();
             $table->string('lokasi');
             // $table->string('description');
         });
